@@ -1,12 +1,10 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-      "./index.html",
-      "./src/**/*.tsx",
-    ],
+export default {
+  content: ['./index.html', './src/**/*.tsx'],
   theme: {
     extend: {},
   },
-  plugins: [require("@tailwindcss/typography"), require("daisyui")],
-}
-
+  plugins: [require('@tailwindcss/typography'), require('daisyui')],
+  daisyui: {
+    themes: ['light', 'coffee'],
+  },
+};
