@@ -1,7 +1,7 @@
 import React from 'react';
+import { ClassName } from 'src/utils';
 
 import { H2 } from '../atomic';
-import { ClassName } from '../global';
 
 type CardProps = ClassName &
   React.PropsWithChildren & {
@@ -9,7 +9,7 @@ type CardProps = ClassName &
   };
 
 export const Card: React.FC<CardProps> = ({ className, title, children }) => (
-  <div className={`card w-96 bg-primary-content ${className}`}>
+  <div className={`card w-full bg-neutral/5 ${className}`}>
     <div className="card-body">
       <H2 className="card-title" content={title} />
       {children}
