@@ -1,7 +1,13 @@
-import {defineConfig} from 'vite'
-import react from '@vitejs/plugin-react'
+import path from 'path';
 
-// https://vitejs.dev/config/
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
+
 export default defineConfig({
-  plugins: [react()]
-})
+  plugins: [react()],
+  resolve: {
+    alias: {
+      src: path.resolve(__dirname, './src'),
+    },
+  },
+});
